some parts separated into componenets
