+++ drizzle/0012_add_profile_capabilities.sql
@@ -1,4 +1,4 @@
-ALTER TABLE "profiles" ADD COLUMN "enabled_capabilities" text[] NOT NULL DEFAULT '{}'::text[];
+ALTER TABLE "profiles" ADD COLUMN "enabled_capabilities" profile_capability[] NOT NULL DEFAULT '{}'::profile_capability[];
 
 -- Add index for faster capability lookups
 CREATE INDEX "profiles_enabled_capabilities_idx" ON "profiles" USING GIN ("enabled_capabilities");
