+++ package.json
@@ -76,6 +76,8 @@
     "swr": "^2.3.0",
     "tailwind-merge": "^2.5.5",
     "tailwindcss-animate": "^1.0.7",
+    "vitest": "^3.0.9",
+    "vitest-mock-extended": "^3.0.1",
     "zod": "^3.24.1",
     "zustand": "^5.0.2"
   },
